Add test client skeleton
