--- client/requirements.txt
@@ -1,1 +1,0 @@
-buffer_buda==0.7.2
